deepseek/requirements.txt
# requirements.txt
z3-solver==4.12.1.0
prometheus-client==0.17.1
docker==6.1.3
kubernetes==26.1.0

# requirements.txt
torch==2.2.0
transformers==4.38.2
fastapi==0.109.0
uvicorn==0.27.0
gunicorn==21.2.0
pyyaml==6.0.1